Entity and Resource resolvers clean up
